Corrgindo erros do sistema, concluindo sistema

+++ FrontConfin/Model/Account.cs
@@ -12,7 +12,7 @@ namespace FrontConfin.Model
         public Guid Id { get; set; }
         public Guid PersonaId { get; set; }
         public string Description { get; set; }
-        public decimal Value { get; set; }
+        public float Value { get; set; }
         public DateTime DueDate { get; set; }
         public DateTime? PaymentDate { get; set; }
         public Situation Situation { get; set; }
